👓 Rename actions
